backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-multipart
httpx
sse-starlette
markdown
pytest
pytest-asyncio
ruff
requests
jira
psycopg2-binary
passlib[bcrypt]
bcrypt<4.1
python-jose[cryptography]
pydantic[email]